LayoutTests/editing/execCommand/4786404-1.html

<html style="font-weight: bold;">
<body style="color: #999;">
<p>This tests for a bug where performing the Remove Format operation on content would apply the document's style to it, not the editable regions default style.  The text below should not be bold, and should be black.</p>
<div id="div" contenteditable="true" style="font-weight: normal; color: black;">foo <i>bar</i> baz</div>

<script>
var div = document.getElementById("div");
div.focus();

document.execCommand("RemoveFormat");
</script>
</body>
